Change strings offset size from 8 bits to 16 bits

+++ src/libruntime/serializer.h
@@ -21,9 +21,9 @@
 
 void serializer_init (void);
 
-uint8_t serializer_dump_strings (const char **, uint8_t);
+uint16_t serializer_dump_strings (const char **, uint8_t);
 
-void serializer_dump_nums (const int32_t *, uint8_t, uint8_t, uint8_t);
+void serializer_dump_nums (const int32_t *, uint8_t, uint16_t, uint8_t);
 
 void serializer_dump_opcode (OPCODE);
